fix fcli dast-scan sc-dast commands fail due to APÏ response body is empty
SC DAST API returns only a 202 response, with an empty body.
Unirest fails to parse the empty body to an objectnode or jsonnode.

The sc-dast cmds now handle this empty body and returns an object with :
- statusCode
- statusText
- message

Also fixed errors when sc-dast cmds had no options/parameters

Refs: #23

+++ fcli-sc-dast/src/main/java/com/fortify/cli/sc_dast/picocli/command/dast_scan/SCDastScanDeleteCommand.java
@@ -35,6 +35,9 @@ import io.micronaut.core.annotation.ReflectiveAccess;
 import kong.unirest.UnirestInstance;
 import lombok.Getter;
 import lombok.SneakyThrows;
+import picocli.CommandLine;
+import picocli.CommandLine.Spec;
+import picocli.CommandLine.Model.CommandSpec;
 import picocli.CommandLine.ArgGroup;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Mixin;
@@ -45,6 +48,8 @@ import picocli.CommandLine.Option;
 @Command(name = "delete", description = "Deletes a DAST scan on ScanCentral DAST")
 @Order(SCDastScanCommandsOrder.DELETE)
 public final class SCDastScanDeleteCommand extends AbstractSCDastUnirestRunnerCommand {
+    @Spec CommandSpec spec;
+
     @ArgGroup(exclusive = false, heading = "Delete scan options:%n", order = 1)
     @Getter private SCDastScanDeleteOptions deleteScanOptions;
 
@@ -58,6 +63,10 @@ public final class SCDastScanDeleteCommand extends AbstractSCDastUnirestRunnerCo
 
     @SneakyThrows
     protected Void runWithUnirest(UnirestInstance unirest) {
+        if(deleteScanOptions == null){
+            throw new CommandLine.ParameterException(spec.commandLine(),
+                    "Error: No parameter found. Provide the required scan-settings identifier.");
+        }
         SCDastScanActionsHandler actionsHandler = new SCDastScanActionsHandler(unirest);
         JsonNode response = actionsHandler.deleteScan(deleteScanOptions.getScanId());
